add nthreads to diffsl creation func

+++ src/vector/faer_serial.rs
@@ -89,13 +89,13 @@ impl<T: Scalar> Vector for Col<T> {
         self.norm_l2()
     }
     fn as_mut_slice(&mut self) -> &mut [Self::T] {
-        unsafe { slice::from_raw_parts_mut(self.as_ptr_mut(), self.len() as usize) }
+        unsafe { slice::from_raw_parts_mut(self.as_ptr_mut(), self.len()) }
     }
     fn as_slice(&self) -> &[Self::T] {
-        unsafe { slice::from_raw_parts(self.as_ptr(), self.len() as usize) }
+        unsafe { slice::from_raw_parts(self.as_ptr(), self.len()) }
     }
     fn copy_from_slice(&mut self, slice: &[Self::T]) {
-        assert_eq!(slice.len(), self.len() as usize, "Vector lengths do not match");
+        assert_eq!(slice.len(), self.len(), "Vector lengths do not match");
         self.iter_mut().zip(slice.iter()).for_each(|(a, b)| *a = *b);
     }
     fn squared_norm(&self, y: &Self, atol: &Self, rtol: Self::T) -> Self::T {
